pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hardstop-agent"
version = "0.3.0"
description = "Local-first event-to-alert risk agent with supply-chain defaults."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.8.0",
    "SQLAlchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.0",
    "feedparser>=6.0.0",
    "requests>=2.31.0",
    "us>=3.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "jsonschema>=4.23.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
hardstop = "hardstop.cli:main"

